+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project changes will be shown here.
 ## [unreleased]
 ### Added
 - Issue 689 - Solarwinds cloud O365 backup API
+- Issue 706 - Create Sales Request - allow attachments
 ### Changed
 - Issue 703 - New calendar
 - Issue 702 - Renewal Report to show users & direct debit